# backend/internal/models/department.go
package models

import (
	"time"

	"github.com/google/uuid"
)

type Department struct {
	ID          uuid.UUID `gorm:"type:uuid;default:gen_random_uuid();primaryKey" json:"id"`
	Name        string    `gorm:"type:varchar(150);not null;uniqueIndex" json:"name"`
	Description string    `gorm:"type:text" json:"description"`
	IsActive    bool      `gorm:"default:true" json:"is_active"`

	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`

	Positions []Position `gorm:"foreignKey:DepartmentID" json:"positions,omitempty"`
}